test/poc/chainlit/requirements.txt
chainlit>=0.7.700
openai>=1.12.0
python-dotenv>=1.0.0
mcp>=0.1.0
langchain
langchain-openai 
openai-agents>=0.0.6 